requirements.txt
# HTTP / scraping
httpx[http2]>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
playwright>=1.44.0

# ORM / migrations
sqlalchemy>=2.0.30
alembic>=1.13.1

# Scheduler
apscheduler>=3.10.4

# Data analysis
pandas>=2.2.2
numpy>=1.26.4

# Utilities
python-dotenv>=1.0.1
tenacity>=8.3.0
fake-useragent>=1.5.1
tqdm>=4.66.4
rich>=13.7.1
loguru>=0.7.2
